refactor: JdbcRepository 구현체 제거
